Added project on preparing features for analysis
This dataset includes information on global diets and COVID morbidity and mortality. The project focuses on effective preparation of data for analysis, but at the end I implement a linear regression model to see if we can effectively predict COVID mortality using information on diet - the answer is that this model needs more work!